Implmented an dynamic programming solution.
Reading a few articles on dynamic programming
to try and get my head around these problems.

This still feels a bit ugly, next step will be to
try and reduce the number of items checked in the
inner loop.
i.e. if we had a sorted collection of best results,
we would only have to loop through until we found
a value form a previous solution less than nums[i],
not all of them up until j.

Worst case this would still be the same,
and the cost of maintaining a sorted collection
may outweigh the gain on a subset of values of i.